Moved types and Effectlevels into GameConstants, got intellij to autogenerate equals method for Position, removed the tests for position as it is no longer needed
